fix tokenizer in api

+++ nemo/lightning/io/connector.py
@@ -18,6 +18,7 @@ from pathlib import Path, PosixPath, WindowsPath
 from typing import Generic, Optional, Tuple, TypeVar
 
 import lightning.pytorch as pl
+import torch
 from filelock import FileLock, Timeout
 from lightning.pytorch.trainer.states import TrainerFn
 
@@ -274,7 +275,7 @@ class ModelConnector(Connector, Generic[SourceT, TargetT]):
             _trainer.strategy.load_model_state_dict(adapter_state, strict=False)
         else:
             _trainer.strategy.load_checkpoint(path)
-
+        model = model.to(torch.bfloat16)
         return model, _trainer
 
     def local_path(self, base_path: Optional[Path] = None) -> Path:
